Fix for issue #745:
IF an NMI or interrupt was taken during single step, the debug cause could be set to trigger or ebreak instead of step in the case wehre a synchronous debug entry cause was in WB but was killed by the NMI or interrupt.

This fix factors in kill_wb when determining if a synchronous debug is pending or not.

Signed-off-by: Oystein Knauserud <Oystein.Knauserud@silabs.com>